+++ main.tf
@@ -13,7 +13,7 @@ locals {
 }
 
 
-data "template_file" "template-nomad-job-presto" {
+data "template_file" "template_nomad_job_presto" {
   template = var.mode == "standalone" ? local.template_standalone : local.template_cluster
 
   vars = {
@@ -54,7 +54,7 @@ data "template_file" "template-nomad-job-presto" {
   }
 }
 
-resource "nomad_job" "nomad-job-presto" {
-  jobspec = data.template_file.template-nomad-job-presto.rendered
+resource "nomad_job" "nomad_job_presto" {
+  jobspec = data.template_file.template_nomad_job_presto.rendered
   detach  = false
 }